COS integration: make prom-api configurable
